+++ src/main/java/io/github/astrapi69/crypt/data/factory/CertFactory.java
@@ -395,24 +395,15 @@ public final class CertFactory
 	 *             is thrown if a security error occur on creation of {@link ContentSigner}
 	 * @throws CertificateException
 	 *             if the conversion is unable to be made
-	 * @throws NoSuchAlgorithmException
-	 *             is thrown if a SecureRandomSpi implementation for the specified algorithm is not
-	 *             available from the specified provider
-	 * @throws NoSuchProviderException
-	 *             is thrown if the specified provider is not registered in the security provider
-	 *             list.
 	 * @throws CertIOException
 	 *             if there is an issue with the new extension value
-	 * @throws InvalidAlgorithmParameterException
-	 *             is thrown if initialization of the cypher object fails
 	 */
-	public static X509Certificate newX509CertificateV3(X509CertificateV3Info certificateInfo)
-		throws OperatorCreationException, CertificateException, NoSuchAlgorithmException,
-		NoSuchProviderException, CertIOException, InvalidAlgorithmParameterException
+	public static X509Certificate newX509CertificateV3(KeyPair keyPair,
+		X509CertificateV3Info certificateInfo)
+		throws OperatorCreationException, CertificateException, CertIOException
 	{
 
 		X509CertificateV1Info v1Info = certificateInfo.getCertificateV1Info();
-		KeyPair keyPair = KeyPairInfo.toKeyPair(v1Info.getKeyPairInfo());
 
 		DistinguishedNameInfo issuer = v1Info.getIssuer();
 		DistinguishedNameInfo subject = v1Info.getSubject();
@@ -455,21 +446,11 @@ public final class CertFactory
 	 *             is thrown if a security error occur on creation of {@link ContentSigner}
 	 * @throws CertificateException
 	 *             if the conversion is unable to be made
-	 * @throws InvalidAlgorithmParameterException
-	 *             is thrown if initialization of the cypher object fails
-	 * @throws NoSuchAlgorithmException
-	 *             is thrown if a SecureRandomSpi implementation for the specified algorithm is not
-	 *             available from the specified provider
-	 * @throws NoSuchProviderException
-	 *             is thrown if the specified provider is not registered in the security provider
-	 *             list.
 	 */
-	public static X509Certificate newX509CertificateV1(X509CertificateV1Info certificateInfo)
-		throws OperatorCreationException, CertificateException, InvalidAlgorithmParameterException,
-		NoSuchAlgorithmException, NoSuchProviderException
+	public static X509Certificate newX509CertificateV1(KeyPair keyPair,
+		X509CertificateV1Info certificateInfo)
+		throws OperatorCreationException, CertificateException
 	{
-
-		KeyPair keyPair = KeyPairInfo.toKeyPair(certificateInfo.getKeyPairInfo());
 		X500Name issuer = new X500Name(certificateInfo.getIssuer().toRepresentableString());
 		BigInteger serial = certificateInfo.getSerial();