+++ src/main/java/io/github/astrapi69/crypt/data/model/X509CertificateV1Info.java
@@ -44,12 +44,6 @@ public class X509CertificateV1Info
 {
 
 	/**
-	 * The key pair information associated with the certificate.
-	 */
-	@NonNull
-	KeyPairInfo keyPairInfo;
-
-	/**
 	 * The distinguished name information of the issuer.
 	 */
 	@NonNull